Changed type of `prev_lines` to `u8` and incremented `prev_lines` when `insert` function is called.
+ `prev_lines` is changed to `u8` to save memory as this could store a list upto 256 and going beyond that is rare as user needs to update netmask and these many devices all using witransfer at same time is also rare.

+++ witransfer/src/utils/terminal.rs
@@ -12,7 +12,7 @@ where
     /// > NOTE: Any modification to `BTreeMap` will not update the terminal directly.
     pub data: BTreeMap<K, String>,
     background_color: Option<Color>,
-    prev_lines: u16,
+    prev_lines: u8,
 }
 
 impl<K> Terminal<K>
